third_party/blink/web_tests/external/wpt/css/css-tables/html5-table-formatting-1-expected.txt

This is a testharness.js-based test.
[PASS] Empty tables can still get a lsyout
[FAIL] Empty tables do not take table-columns into account
  assert_equals: expected 50 but got 200
[FAIL] Empty tables do not take table-rows into account
  assert_equals: expected 50 but got 100
[PASS] Table-columns are taken into account after missing cells are generated (empty line)
[PASS] Table-columns are taken into account after missing cells are generated (partially empty line)
[PASS] Table-columns are taken into account after missing cells are generated (non-empty line)
Harness: the test ran to completion.
